requirements.txt
Flask
requests
deep-translator
gunicorn
python-dotenv